Modifies TypeIndex macro to require that a reference to the originating code is pushed down into it. A limitation with the macro system means that nested macros don't report from where an error originates, but the macro above. By passing a reference in, when TypeIndex is unable to index a data-structure, the developer building on-top of dust can work out where in their code the error is sourced. This change has ramified into other areas of the code where the TypeIndex macro has been used.
rakefile.rb is modified to provide the rake task "haxelib:local", which properly installs dust into haxelib from a local git clone. Care should be taken because this results in two copies of dust. Changes to dust are therefore not recommended. This is temporary until a stable release makes it redundant.